+++ app/controllers/application_controller.rb
@@ -1,6 +1,7 @@
 class ApplicationController < ActionController::Base
   before_action :authenticate_user!
   # before_action :configure_permitted_parameters, if: :devise_controller?
+  before_action :configure_account_update_params, if: :devise_controller?
 
   def set_search_hostel
     if params[:q] != nil
@@ -16,9 +17,35 @@ class ApplicationController < ActionController::Base
       # @hostels = @q.result(distinct: true)
   end
 
-  # protected
+  protected
   # def configure_permitted_parameters
   #   devise_parameter_sanitizer.permit(:sign_up, keys: [:role, :name])
   # end
 
+  def configure_account_update_params
+    devise_parameter_sanitizer.permit(:account_update, keys: [
+      :name,
+      host_attributes: [
+        :name,
+        :phone_number,
+        :profile,
+        :image,
+        :country,
+        :birthday,
+        :user_id
+        ],
+      guest_attributes: [
+        :address,
+        :phone_number,
+        :gender,
+        :country,
+        :latitude,
+        :longitude,
+        :image,
+        :birthday,
+        :user_id
+        ]
+      ]
+    )
+  end
 end
