+++ app/controllers/users_controller.rb
@@ -1,0 +1,6 @@
+class UsersController < ApplicationController
+  def show
+    @user = current_user
+    # binding.irb
+  end
+end
